aiml/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.1.0
scikit-learn>=1.3.0
numpy>=1.26.0
pandas>=2.1.0
pyyaml>=6.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
